example_versatile/requirements-current.txt
django-versatileimagefield>=1.11.0
django~=1.11
